+++ lib/td_auth/auth/pipeline/current_resource.ex
@@ -1,18 +1,17 @@
 defmodule TdAuth.Auth.Pipeline.CurrentResource do
-    @moduledoc """
-    A plug to read the current resource from Guardian and assign it to the :current_resource
-    key in the connection.
-    """
+  @moduledoc """
+  A plug to read the claims from Guardian and assign them to the :current_resource
+  key in the connection.
+  """
 
-    use Plug.Builder
-    alias Guardian.Plug, as: GuardianPlug
+  use Plug.Builder
 
-    plug(:current_resource)
+  plug(:current_resource)
 
-    def init(opts), do: opts
+  def init(opts), do: opts
 
-    def current_resource(conn, _opts) do
-      current_resource = GuardianPlug.current_resource(conn)
-      assign(conn, :current_resource, current_resource)
-    end
+  def current_resource(conn, _opts) do
+    claims = Guardian.Plug.current_resource(conn)
+    assign(conn, :current_resource, claims)
   end
+end