+++ lib/td_auth_web/controllers/permission_group_controller.ex
@@ -39,9 +39,9 @@ defmodule TdAuthWeb.PermissionGroupController do
   end
 
   def create(conn, %{"permission_group" => permission_group_params}) do
-    current_resource = conn.assigns[:current_resource]
+    claims = conn.assigns[:current_resource]
 
-    with {:can, true} <- {:can, current_resource.is_admin},
+    with {:can, true} <- {:can, claims.role === "admin"},
          {:ok, %PermissionGroup{} = permission_group} <-
            Permissions.create_permission_group(permission_group_params) do
       conn
@@ -85,9 +85,9 @@ defmodule TdAuthWeb.PermissionGroupController do
   end
 
   def update(conn, %{"id" => id, "permission_group" => permission_group_params}) do
-    current_resource = conn.assigns[:current_resource]
+    claims = conn.assigns[:current_resource]
 
-    with {:can, true} <- {:can, current_resource.is_admin},
+    with {:can, true} <- {:can, claims.is_admin},
          permission_group <- Permissions.get_permission_group!(id),
          {:ok, %PermissionGroup{} = permission_group} <-
            Permissions.update_permission_group(permission_group, permission_group_params) do
@@ -109,9 +109,9 @@ defmodule TdAuthWeb.PermissionGroupController do
   end
 
   def delete(conn, %{"id" => id}) do
-    current_resource = conn.assigns[:current_resource]
+    claims = conn.assigns[:current_resource]
 
-    with {:can, true} <- {:can, current_resource.is_admin},
+    with {:can, true} <- {:can, claims.is_admin},
          permission_group <- Permissions.get_permission_group!(id),
          {:ok, %PermissionGroup{}} <- Permissions.delete_permission_group(permission_group) do
       send_resp(conn, :no_content, "")